+++ SteamStorage/ControlElements/RemainElements/SellRemainElement.xaml.cs
@@ -45,7 +45,7 @@ namespace SteamStorage.ControlElements
             try
             {
                 string Url = GeneralMethods.GetSkin(RemainElementFull.IdSkin).Url;
-                int count = Math.Min(Convert.ToInt32(Count.Text), RemainElementFull.Count);
+                long count = Math.Min(Convert.ToInt32(Count.Text), RemainElementFull.Count);
                 Exception ex = ArchiveMethods.AddNewArchiveElement(Url, count,
                     Convert.ToDouble(CostPurchase.Text.Replace('.', ',')),
                     Convert.ToDouble(CostSold.Text.Replace('.', ',')),
